Restructurize auth modules and remove reduntant one
