+++ core/che-core-api-core/src/main/java/org/eclipse/che/api/core/rest/AuthenticationExceptionMapper.java
@@ -9,13 +9,14 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.che.commons.auth;
+package org.eclipse.che.api.core.rest;
 
 import javax.inject.Singleton;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
+import org.eclipse.che.api.core.AuthenticationException;
 import org.eclipse.che.dto.server.DtoFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
