+++ core/commons/pom.xml
@@ -25,7 +25,6 @@
     <name>Che Core :: Commons :: Parent</name>
     <modules>
         <module>che-core-commons-annotations</module>
-        <module>che-core-commons-auth</module>
         <module>che-core-commons-lang</module>
         <module>che-core-commons-logback</module>
         <module>che-core-commons-inject</module>
