+++ pom.xml
@@ -364,11 +364,6 @@
             </dependency>
             <dependency>
                 <groupId>org.eclipse.che.core</groupId>
-                <artifactId>che-core-commons-auth</artifactId>
-                <version>${che.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>org.eclipse.che.core</groupId>
                 <artifactId>che-core-commons-inject</artifactId>
                 <version>${che.version}</version>
             </dependency>
